make RemoveUnused enhancement support Scala 2.12 & 2.13
