+++ scalafix-tests/input/src/main/scala-2.13/test/RemoveUnusedParams.scala
@@ -1,0 +1,9 @@
+/*
+rule = RemoveUnused
+ */
+package test
+
+object UnusedParams {
+  def g(x: String => Unit): Unit = ???
+  g{implicit string => println("g")}
+}
