+++ scalafix-tests/input/src/main/scala-2.13/test/RemoveUnusedPatternVars.scala
@@ -57,6 +57,18 @@ object ob {
       }
   }
 
+  def pf(x: PartialFunction[Any, Unit]): Unit = ???
+  case class A(a: Int)
+  pf{
+    case string: String => ???
+    case (i: Int) => ???
+    case (a: Int, b) => println(b)
+    case a@A(v) => ???
+    case x :: (y1, y2) :: Nil => println(x)
+    case (zz) => ???
+  }
+  try ??? catch {case e: Exception => ???}
+
   def a: Unit = {
     example match {
       case AB(_, _) => println("Not used, good")
