+++ scalafix-tests/input/src/main/scala-2/test/RemoveUnusedTerms.scala
@@ -19,18 +19,4 @@ object RemoveUnusedTerms {
   val dd = 0
   def f(x: Int) = "unused"
   private def ff(x: Int) = "unused"
-
-  def g(x: String => Unit): Unit = ???
-  g{implicit string => println("g")}
-  def pf(x: PartialFunction[Any, Unit]): Unit = ???
-  case class A(a: Int)
-  pf{
-    case string: String => ???
-    case (i: Int) => ???
-    case (a: Int, b) => println(b)
-    case a@A(v) => ???
-    case x :: (y1, y2) :: Nil => println(x)
-    case (zz) => ???
-  }
-  try ??? catch {case e: Exception => ???}
 }
