+++ scalafix-tests/output/src/main/scala-2.12/test/RemoveUnusedParams.scala
@@ -1,0 +1,6 @@
+package test
+
+object UnusedParams {
+  def g(x: String => Unit): Unit = ???
+  g{_ => println("g")}
+}
