+++ scalafix-tests/output/src/main/scala-2.12/test/RemoveUnusedPatternVars.scala
@@ -53,4 +53,16 @@ object Unusedmatchargs {
         }
       }
   }
+
+  def pf(x: PartialFunction[Any, Unit]): Unit = ???
+  case class A(a: Int)
+  pf{
+    case _: String => ???
+    case (_: Int) => ???
+    case (_: Int, b) => println(b)
+    case A(_) => ???
+    case x :: (_, _) :: Nil => println(x)
+    case _ => ???
+  }
+  try ??? catch {case _: Exception => ???}
 }
