+++ scalafix-tests/output/src/main/scala-2/test/RemoveUnusedTerms.scala
@@ -16,18 +16,4 @@ object RemoveUnusedTerms {
   val dd = 0
   def f(x: Int) = "unused"
   
-
-  def g(x: String => Unit): Unit = ???
-  g{_ => println("g")}
-  def pf(x: PartialFunction[Any, Unit]): Unit = ???
-  case class A(a: Int)
-  pf{
-    case _: String => ???
-    case (_: Int) => ???
-    case (_: Int, b) => println(b)
-    case A(_) => ???
-    case x :: (_, _) :: Nil => println(x)
-    case _ => ???
-  }
-  try ??? catch {case _: Exception => ???}
 }
